% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused_formals.R
\name{unused_formals}
\alias{unused_formals}
\title{Is every function argument used in its body?}
\usage{
unused_formals(fn, ns = NULL)
}
\arguments{
\item{fn}{A function.}

\item{ns}{If \code{fn} is a character vector, the namespace to which \code{fn} belongs.}
}
\value{
The arguments of \code{fn} that are not used or if all are indeed used \code{NULL} invisibly.
}
\description{
Is every function argument used in its body?
}
\examples{
add3 <- function(x, y, z = 0) {
  x + y
}

unused_formals(add3)
}
