% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formalPackageCoverage.R
\name{formalPackageCoverage}
\alias{formalPackageCoverage}
\title{Formal coverage of package}
\usage{
formalPackageCoverage(pkg = ".")
}
\arguments{
\item{pkg}{Top level of package.}
}
\value{
If no formals uncovered, returns \code{NULL}.
}
\description{
Formal coverage of package
}
